+++ src/components/Notification/notificationChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: ["Removes deprecated `noMargin` prop."],
+  },
+  {
     date: "2025-04-24",
     version: "3.6.1",
     type: "Update",